requirements.txt
Flask
numpy
scikit-learn
requests
python-dotenv
gunicorn